+++ src/libcolorscreen/render-tile.h
@@ -50,7 +50,7 @@ bool render_img_normal(render_type_parameters rtparam,
     progress->pop ();
   if (progress)
     progress->set_task ("rendering", height);
-#pragma omp parallel for default(none) shared(progress,pixels,render,pixelbytes,rowstride,height, width,step,yoffset,xoffset) if (width * height > render.openmp_size ())
+#pragma omp parallel for default(none) shared(progress,pixels,render,pixelbytes,rowstride,height, width,step,yoffset,xoffset) if (width * (size_t)height > render.openmp_size ())
   for (int y = 0; y < height; y++)
     {
       coord_t py = (y + yoffset) * step;
@@ -96,7 +96,7 @@ bool render_img_downscale(render_type_parameters rtparam,
   render.get_color_data (data, xoffset * step, yoffset * step, width, height, step, progress);
   if (progress)
     progress->set_task ("rendering", height);
-#pragma omp parallel for default(none) shared(progress,pixels,render,pixelbytes,rowstride,height, width,step,yoffset,xoffset,data) if (width * height > render.openmp_size ())
+#pragma omp parallel for default(none) shared(progress,pixels,render,pixelbytes,rowstride,height, width,step,yoffset,xoffset,data) if (width * (size_t)height > render.openmp_size ())
   for (int y = 0; y < height; y++)
     {
       if (!progress || !progress->cancel_requested ())
@@ -141,7 +141,7 @@ bool render_img_gray_downscale(render_type_parameters rtparam,
   render.get_gray_data (data, xoffset * step, yoffset * step, width, height, step, progress);
   if (progress)
     progress->set_task ("rendering", height);
-#pragma omp parallel for default(none) shared(progress,pixels,render,pixelbytes,rowstride,height, width,step,yoffset,xoffset,data) if (width * height > render.openmp_size ())
+#pragma omp parallel for default(none) shared(progress,pixels,render,pixelbytes,rowstride,height, width,step,yoffset,xoffset,data) if (width * (size_t)height > render.openmp_size ())
   for (int y = 0; y < height; y++)
     {
       if (!progress || !progress->cancel_requested ())
@@ -325,7 +325,7 @@ void render_stitched(RP &rtparam, P &outer_param,
 	    if (progress)
 	      progress->push ();
 	    renders[iy * stitch.params.width + ix]  = init_render (rtparam, rparam2, *img.stitch->images[iy][ix].img, img.stitch->images[iy][ix].param, outer_param, progress);
-	    openmp_size = renders[iy * stitch.params.width + ix]->openmp_size ();
+	    openmp_size = renders[iy * (size_t)stitch.params.width + ix]->openmp_size ();
 	    if (progress)
 	      {
 		progress->pop ();
